refactor: 회원 전용 API > member 디렉터리로 이동
